refactor(perps): unify Lite and Pro market headers
Extract shared PerpsMarketHeader and PerpsMarketSummary from the Pro
implementation so both market detail screens share one canonical header
with scroll-linked price crossfade. Lite keeps its own endAccessory
actions, restores safe-area top inset, and uses a content-hugging large
price row. Order book keeps the compact inline header only.
