+++ app/components/UI/Perps/Views/PerpsMarketDetailsView/PerpsMarketDetailsView.tsx
@@ -31,7 +31,15 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import { Linking, RefreshControl, ScrollView, View } from 'react-native';
+import {
+  Linking,
+  RefreshControl,
+  View,
+  type LayoutChangeEvent,
+  type NativeScrollEvent,
+} from 'react-native';
+import Animated, { useAnimatedScrollHandler } from 'react-native-reanimated';
+import { scheduleOnRN } from 'react-native-worklets';
 import {
   CandlePeriod,
   PerpsMode,
@@ -71,8 +79,10 @@ import {
   PerpsTutorialSelectorsIDs,
   PerpsCompactOrderRowSelectorsIDs,
 } from '../../Perps.testIds';
-import LivePriceHeader from '../../components/LivePriceDisplay/LivePriceHeader';
-import PerpsMarketInlineHeader from '../../components/PerpsMarketInlineHeader';
+import PerpsMarketHeader, {
+  createLiteMarketHeaderTestIDs,
+} from '../../components/PerpsMarketHeader';
+import PerpsMarketSummary from '../../components/PerpsMarketSummary';
 import PerpsModeToggle from '../../components/PerpsModeToggle';
 import { showPerpsModeFlash } from '../../utils/perpsModeFlash';
 import PerpsMarketAboutSection from '../../components/PerpsMarketAboutSection';
@@ -296,7 +306,7 @@ const PerpsMarketDetailsView: React.FC<PerpsMarketDetailsViewProps> = () => {
   // This prevents stale closure issues where the captured position is outdated
   // Initialized to null, will be updated via useEffect when existingPosition is available
   const currentPositionRef = useRef<Position | null>(null);
-  const scrollViewRef = useRef<ScrollView>(null);
+  const scrollViewRef = useRef<Animated.ScrollView>(null);
 
   const isEligible = useSelector(selectPerpsEligibility);
 
@@ -391,10 +401,53 @@ const PerpsMarketDetailsView: React.FC<PerpsMarketDetailsViewProps> = () => {
 
   const {
     scrollY: scrollYShared,
-    onScroll,
     setTitleSectionHeight,
     titleSectionHeightSv,
   } = useHeaderStandardAnimated();
+
+  const handleMarketSummaryLayout = useCallback(
+    (event: LayoutChangeEvent) => {
+      const { height } = event.nativeEvent.layout;
+      if (height > 0) {
+        setTitleSectionHeight(height);
+      }
+    },
+    [setTitleSectionHeight],
+  );
+
+  const handleAboutScrollEvent = useCallback(
+    (scrollY: number, viewportHeight: number) => {
+      handleAboutScroll({
+        nativeEvent: {
+          contentOffset: { x: 0, y: scrollY },
+          layoutMeasurement: { width: 0, height: viewportHeight },
+        } as NativeScrollEvent,
+      });
+    },
+    [handleAboutScroll],
+  );
+
+  const handleAboutScrollEventRef = useRef(handleAboutScrollEvent);
+  handleAboutScrollEventRef.current = handleAboutScrollEvent;
+  const scheduleAboutScrollOnRN = useCallback(
+    (scrollY: number, viewportHeight: number) => {
+      handleAboutScrollEventRef.current(scrollY, viewportHeight);
+    },
+    [],
+  );
+
+  const scrollHandler = useAnimatedScrollHandler({
+    onScroll: (event) => {
+      'worklet';
+      // eslint-disable-next-line react-compiler/react-compiler -- Reanimated shared values are intentionally mutated from worklets.
+      scrollYShared.value = event.contentOffset.y;
+      scheduleOnRN(
+        scheduleAboutScrollOnRN,
+        event.contentOffset.y,
+        event.layoutMeasurement.height,
+      );
+    },
+  });
 
   // Get persisted candle period preference from Redux store
   const selectedCandlePeriod = useSelector(
@@ -886,6 +939,48 @@ const PerpsMarketDetailsView: React.FC<PerpsMarketDetailsViewProps> = () => {
       [PERPS_EVENT_PROPERTY.FAVORITES_COUNT]: watchlistCount,
     });
   }, [market, isWatchlist, track, showToast, PerpsToastOptions]);
+
+  const liteHeaderEndAccessory = useMemo(() => {
+    const favoriteButton = (
+      <ButtonIcon
+        iconName={isWatchlist ? IconName.StarFilled : IconName.Star}
+        size={ButtonIconSize.Md}
+        onPress={handleWatchlistPress}
+        testID={PerpsMarketHeaderSelectorsIDs.FAVORITE_BUTTON}
+        accessibilityLabel={strings(
+          isWatchlist
+            ? 'perps.market_details.remove_from_watchlist'
+            : 'perps.market_details.add_to_watchlist',
+        )}
+      />
+    );
+
+    if (!isPerpsProModeEnabled) {
+      return favoriteButton;
+    }
+
+    return (
+      <Box
+        flexDirection={BoxFlexDirection.Row}
+        alignItems={BoxAlignItems.Center}
+        gap={2}
+      >
+        {favoriteButton}
+        <PerpsModeToggle
+          mode={perpsMode}
+          variant="active"
+          onChange={handlePerpsModeChange}
+          source={PERPS_EVENT_VALUE.SOURCE.PERP_ASSET_SCREEN}
+        />
+      </Box>
+    );
+  }, [
+    isWatchlist,
+    handleWatchlistPress,
+    isPerpsProModeEnabled,
+    perpsMode,
+    handlePerpsModeChange,
+  ]);
 
   const handleMarketListPress = useCallback(() => {
     if (!market) return;
@@ -1576,95 +1671,55 @@ const PerpsMarketDetailsView: React.FC<PerpsMarketDetailsViewProps> = () => {
   return (
     <SafeAreaView
       style={styles.mainContainer}
-      edges={['bottom', 'left', 'right']}
+      edges={['top', 'bottom', 'left', 'right']}
       testID={PerpsMarketDetailsViewSelectorsIDs.CONTAINER}
     >
-      {/* Only the market identity (icon + name + ticker + leverage) is a tap
-          target that opens the market list; it renders as a content-hugging
-          box inside the header. */}
-      <PerpsMarketInlineHeader
+      <PerpsMarketHeader
         market={market}
-        currentPrice={syncedChartCurrentPrice}
+        testIDs={createLiteMarketHeaderTestIDs()}
         onBackPress={handleBackPress}
-        onFavoritePress={handleWatchlistPress}
         onIdentityPress={handleMarketListPress}
-        isFavorite={isWatchlist}
-        useDetailLayout
-        testID={PerpsMarketDetailsViewSelectorsIDs.HEADER}
-        // Pro mode: keep the watchlist star and append a read-only pill showing
-        // the active mode. `endAccessory` replaces the header's default end
-        // icons, so the star is re-composed here alongside the pill.
-        endAccessory={
-          isPerpsProModeEnabled ? (
-            <Box
-              flexDirection={BoxFlexDirection.Row}
-              alignItems={BoxAlignItems.Center}
-              gap={2}
-            >
-              <ButtonIcon
-                iconName={isWatchlist ? IconName.StarFilled : IconName.Star}
-                size={ButtonIconSize.Md}
-                onPress={handleWatchlistPress}
-                testID={PerpsMarketHeaderSelectorsIDs.FAVORITE_BUTTON}
-              />
-              <PerpsModeToggle
-                mode={perpsMode}
-                variant="active"
-                onChange={handlePerpsModeChange}
-                source={PERPS_EVENT_VALUE.SOURCE.PERP_ASSET_SCREEN}
-              />
-            </Box>
-          ) : undefined
-        }
+        scrollY={scrollYShared}
+        priceSectionHeight={titleSectionHeightSv}
+        endAccessory={liteHeaderEndAccessory}
       />
 
       <View style={styles.scrollableContentContainer}>
-        <ScrollView
+        <Animated.ScrollView
           ref={scrollViewRef}
           style={styles.mainContentScrollView}
           contentContainerStyle={styles.scrollViewContent}
           showsVerticalScrollIndicator={false}
-          onScroll={handleAboutScroll}
+          onScroll={scrollHandler}
           scrollEventThrottle={16}
           testID={PerpsMarketDetailsViewSelectorsIDs.SCROLL_VIEW}
           refreshControl={
             <RefreshControl refreshing={refreshing} onRefresh={handleRefresh} />
           }
         >
-          {/* Below header: live price + 24h change and fullscreen chart button.
-              Rendered inside the ScrollView (not the header) so it scrolls with
-              the content instead of staying sticky. */}
-          <Box
-            flexDirection={BoxFlexDirection.Row}
-            alignItems={BoxAlignItems.Center}
-            justifyContent={BoxJustifyContent.Between}
-            gap={2}
-            twClassName="px-4 pb-2"
+          <PerpsMarketSummary
+            symbol={market.symbol}
+            currentPrice={syncedChartCurrentPrice}
             testID={PerpsMarketDetailsViewSelectorsIDs.MARKET_SUMMARY}
-          >
-            {/* Flexible wrapper lets the price shrink; the button stays fixed. */}
-            <Box twClassName="flex-1">
-              <LivePriceHeader
-                symbol={market.symbol}
-                testIDPrice={PerpsMarketHeaderSelectorsIDs.PRICE}
-                testIDChange={PerpsMarketHeaderSelectorsIDs.PRICE_CHANGE}
-                currentPrice={syncedChartCurrentPrice}
-                size="large"
+            testIDPrice={PerpsMarketHeaderSelectorsIDs.PRICE}
+            testIDChange={PerpsMarketHeaderSelectorsIDs.PRICE_CHANGE}
+            size="large"
+            onLayout={handleMarketSummaryLayout}
+            endAccessory={
+              <ButtonIcon
+                iconName={IconName.Expand}
+                size={ButtonIconSize.Md}
+                onPress={handleFullscreenChartOpen}
+                style={styles.marketSummaryFullscreenButton}
+                testID={
+                  PerpsMarketDetailsViewSelectorsIDs.FULLSCREEN_CHART_BUTTON
+                }
+                accessibilityLabel={strings(
+                  'perps.market_details.fullscreen_chart',
+                )}
               />
-            </Box>
-            <ButtonIcon
-              iconName={IconName.Expand}
-              size={ButtonIconSize.Md}
-              onPress={handleFullscreenChartOpen}
-              style={styles.marketSummaryFullscreenButton}
-              testID={
-                PerpsMarketDetailsViewSelectorsIDs.FULLSCREEN_CHART_BUTTON
-              }
-              accessibilityLabel={strings(
-                'perps.market_details.fullscreen_chart',
-              )}
-            />
-          </Box>
+            }
+          />
 
           {/* TradingView Chart Section */}
           <View style={[styles.section, styles.chartSection]}>
@@ -1832,7 +1887,7 @@ const PerpsMarketDetailsView: React.FC<PerpsMarketDetailsViewProps> = () => {
               </Text>
             </Text>
           </View>
-        </ScrollView>
+        </Animated.ScrollView>
       </View>
 
       {/* Fixed Actions Footer */}
