+++ app/components/UI/Perps/components/PerpsMarketSummary/index.ts
@@ -1,0 +1,2 @@
+export { default, PRICE_SECTION_HEIGHT } from './PerpsMarketSummary';
+export type { PerpsMarketSummaryProps } from './PerpsMarketSummary';
